+++ fontforgeexe/uiutil.c
@@ -39,106 +39,6 @@ extern GBox _ggadget_Default_Box;
 #define ACTIVE_BORDER   (_ggadget_Default_Box.active_border)
 #define MAIN_FOREGROUND (_ggadget_Default_Box.main_foreground)
 
-
-static char* SupportedLocale(const char *locale, const char *fullspec, const char *filename) {
-/* If there's additional help files written for other languages, then check */
-/* to see if this local matches the additional help message language. If so */
-/* then report back that there's another language available to use for help */
-/* NOTE: If Docs are not maintained very well, maybe comment-out lang here. */
-    int i;
-    /* list languages in specific to generic order, ie: en_CA, en_GB, en... */
-    static char *supported[] = { "de","ja", NULL }; /* other html lang list */
-
-    for ( i=0; supported[i]!=NULL; ++i ) {
-        if ( strcmp(locale,supported[i])==0 ) {
-            char *pt = strrchr(filename, '/');
-            return smprintf("%s/old/%s/%s", fullspec, supported[i], pt ? pt : filename);
-        }
-    }
-    return NULL;
-}
-
-static char* CheckSupportedLocale(const char *fullspec, const char *filename) {
-/* Add Browser HELP for this local if there's more html docs for this local */
-
-    /* KANOU has provided a japanese translation of the docs */
-    /* Edward Lee is working on traditional chinese docs */
-    const char *loc = getenv("LC_ALL");
-    char buffer[40], *pt;
-
-    if ( loc==NULL ) loc = getenv("LC_CTYPE");
-    if ( loc==NULL ) loc = getenv("LANG");
-    if ( loc==NULL ) loc = getenv("LC_MESSAGES");
-    if ( loc==NULL )
-        return NULL;
-
-    /* first, try checking entire string */
-    strncpy(buffer,loc,sizeof(buffer));
-    buffer[sizeof(buffer)-1] = '\0';
-    pt = SupportedLocale(buffer, fullspec, filename);
-    if (pt) {
-        return pt;
-    }
-
-    /* parse possible suffixes, such as .UTF-8, then try again */
-    if ( (pt=strchr(buffer,'.'))!=NULL ) {
-        *pt = '\0';
-        pt = SupportedLocale(buffer, fullspec, filename);
-        if (pt) {
-            return pt;
-        }
-    }
-
-    /* parse possible suffixes such as _CA, _GB, and try again */
-    if ( (pt=strchr(buffer,'_'))!=NULL ) {
-        *pt = '\0';
-        return SupportedLocale(buffer, fullspec, filename);
-    }
-
-    return NULL;
-}
-
-void help(const char *file, const char *section) {
-    if (!file) {
-        return;
-    } else if (strstr(file, "://")) {
-        g_app_info_launch_default_for_uri(file, NULL, NULL);
-        return;
-    } else if (!section) {
-        section = "";
-    }
-
-    bool launched = false;
-    const char *help = getHelpDir();
-    if (help) {
-        char *path = CheckSupportedLocale(help, file);
-        if (!path) {
-            path = smprintf("%s/%s", help, file);
-            if (!path) {
-                return;
-            }
-        }
-
-        GFile *gfile = g_file_new_for_path(path);
-        free(path);
-
-        if (g_file_query_exists(gfile, NULL)) {
-            gchar* uri = g_file_get_uri(gfile);
-            path = smprintf("%s%s", uri, section);
-            launched = g_app_info_launch_default_for_uri(path, NULL, NULL);
-            g_free(uri);
-            free(path);
-        }
-        g_object_unref(gfile);
-    }
-
-    if (!launched) {
-        char *path = smprintf("https://fontforge.org/docs/%s%s", file, section);
-        g_app_info_launch_default_for_uri(path, NULL, NULL);
-        free(path);
-    }
-}
-
 static void UI_IError(const char *format,...) {
     va_list ap;
     char buffer[300];
